Support Mul(Add(...), b)
