+++ test/multests.jl
@@ -686,96 +686,6 @@ import Base.Broadcast: materialize, materialize!
         M = Mul(A, Mul(B, C))
         @test ndims(M) == ndims(typeof(M)) == 2
         @test eltype(M) == Float64
-        @test_broken all(copyto!(R1, M) .=== A*(B*C) .=== (R2 .= M))
-    end
-end
-
-
-@testset "Add" begin
-    @testset "gemv Float64" begin
-        for A in (AddArray(randn(5,5), randn(5,5)),
-                  AddArray(randn(5,5), view(randn(9, 5), 1:2:9, :))),
-            b in (randn(5), view(randn(5),:), view(randn(5),1:5), view(randn(9),1:2:9))
-
-            Ã = copy(A)
-            c = similar(b)
-
-            c .= Mul(A,b)
-            @test c ≈ Ã*b ≈ BLAS.gemv!('N', 1.0, Ã, b, 0.0, similar(c))
-
-            copyto!(c, Mul(A,b))
-            @test c ≈ Ã*b ≈ BLAS.gemv!('N', 1.0, Ã, b, 0.0, similar(c))
-
-            b̃ = copy(b)
-            copyto!(b̃, Mul(A,b̃))
-            @test c ≈ b̃
-
-            c .= 2.0 .* Mul(A,b)
-            @test c ≈ BLAS.gemv!('N', 2.0, Ã, b, 0.0, similar(c))
-
-            c = copy(b)
-            c .= Mul(A,b) .+ c
-            @test c ≈ BLAS.gemv!('N', 1.0, Ã, b, 1.0, copy(b))
-
-            c = copy(b)
-            c .= Mul(A,b) .+ 2.0 .* c
-            @test c ≈ BLAS.gemv!('N', 1.0, Ã, b, 2.0, copy(b))
-
-            c = copy(b)
-            c .= 2.0 .* Mul(A,b) .+ c
-            @test c ≈ BLAS.gemv!('N', 2.0, Ã, b, 1.0, copy(b))
-
-            c = copy(b)
-            c .= 3.0 .* Mul(A,b) .+ 2.0 .* c
-            @test c ≈ BLAS.gemv!('N', 3.0, Ã, b, 2.0, copy(b))
-
-            d = similar(c)
-            c = copy(b)
-            d .= 3.0 .* Mul(A,b) .+ 2.0 .* c
-            @test d ≈ BLAS.gemv!('N', 3.0, Ã, b, 2.0, copy(b))
-        end
-    end
-
-    @testset "gemm" begin
-        for A in (AddArray(randn(5,5), randn(5,5)),
-                  AddArray(randn(5,5), view(randn(9, 5), 1:2:9, :))),
-            B in (randn(5,5), view(randn(5,5),:,:), view(randn(5,5),1:5,:),
-                  view(randn(5,5),1:5,1:5), view(randn(5,5),:,1:5))
-
-            Ã = copy(A)
-            C = similar(B)
-
-            C .= Mul(A,B)
-            @test C ≈ BLAS.gemm!('N', 'N', 1.0, Ã, B, 0.0, similar(C))
-
-            B .= Mul(A,B)
-            @test C ≈ B
-
-            C .= 2.0 .* Mul(A,B)
-            @test C ≈ BLAS.gemm!('N', 'N', 2.0, Ã, B, 0.0, similar(C))
-
-            C = copy(B)
-            C .= Mul(A,B) .+ C
-            @test C ≈ BLAS.gemm!('N', 'N', 1.0, Ã, B, 1.0, copy(B))
-
-
-            C = copy(B)
-            C .= Mul(A,B) .+ 2.0 .* C
-            @test C ≈ BLAS.gemm!('N', 'N', 1.0, Ã, B, 2.0, copy(B))
-
-            C = copy(B)
-            C .= 2.0 .* Mul(A,B) .+ C
-            @test C ≈ BLAS.gemm!('N', 'N', 2.0, Ã, B, 1.0, copy(B))
-
-
-            C = copy(B)
-            C .= 3.0 .* Mul(A,B) .+ 2.0 .* C
-            @test C ≈ BLAS.gemm!('N', 'N', 3.0, Ã, B, 2.0, copy(B))
-
-            d = similar(C)
-            C = copy(B)
-            d .= 3.0 .* Mul(A,B) .+ 2.0 .* C
-            @test d ≈ BLAS.gemm!('N', 'N', 3.0, Ã, B, 2.0, copy(B))
-        end
+        @test all(copyto!(R1, M) .=== A*(B*C) .=== (R2 .= M))
     end
 end
