+++ test/runtests.jl
@@ -4,6 +4,7 @@ include("memorylayouttests.jl")
 include("applytests.jl")
 include("multests.jl")
 include("ldivtests.jl")
+include("addtests.jl")
 include("setoptests.jl")
